requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-bootstrap3==11.1.0
django-ckeditor==5.7.1
django-js-asset==1.2.2
gunicorn==19.9.0
psycopg2-binary==2.8.3
python-dotenv==0.10.3
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.3
